% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise.R
\name{chronik_by_source_place}
\alias{chronik_by_source_place}
\title{summarising all data files into handy formats for visualization}
\usage{
chronik_by_source_place()
}
\arguments{
\item{input, output, session}{Internal parameters for {shiny}. 
DO NOT REMOVE.}
}
\description{
summarising all data files into handy formats for visualization
}
